IRISVOICE/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
pydantic>=2.10.0
aiofiles>=24.0.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
requests>=2.31.0

# Audio/Voice Processing (Native Audio Phase 1)
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.11.0
pvporcupine>=3.0.0
silero-vad>=4.0.0
sounddevice>=0.5.0
# PyTorch with CUDA support (GPU)
# For GPU users: install with CUDA using:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU-only users: keep as is
torch>=2.0.0
torchaudio>=2.0.0
liquid-audio>=0.1.0

# Vision & Automation (Phase 2)
mss>=9.0.0
Pillow>=10.0.0
pyautogui>=0.9.54
anthropic>=0.18.0
openai>=1.0.0